feat[gpu]: dict take kernel

+++ vortex-gpu/src/take.rs
@@ -161,7 +161,7 @@ mod tests {
         const LEN: usize = 1024 * 8;
         let values: PrimitiveArray = (0i64..1024).map(|x| (x + 2) % 1024).collect();
         let codes: PrimitiveArray = (0..LEN)
-            .map(|x| (u8::try_from(((x + 1) % 255) as u8).vortex_expect("")))
+            .map(|x| (u8::try_from(((x + 1) % 255)).vortex_expect("")))
             .collect();
         let dict = DictArray::try_new(codes.into_array(), values.into_array()).unwrap();
